Making total score a double

+++ src/main/java/com/example/demo/HomePage.java
@@ -50,7 +50,7 @@ public class HomePage {
 	
 	
 	@GetMapping("/results")
-	public int surveyResults() {
+	public double surveyResults() {
 		return surveyManageService.calculateScore(surveyQuestionsResults);
 	}
